+++ src/test/java/net/openhft/chronicle/wire/MyClass2.java
@@ -1,0 +1,7 @@
+package net.openhft.chronicle.wire;
+
+public class MyClass2 extends SelfDescribingMarshallable {
+
+    MyClass3 myClass;
+
+}
