+++ src/test/java/net/openhft/chronicle/wire/StringConsumerMarshallableOutTest.java
@@ -16,10 +16,11 @@
  * limitations under the License.
  */
 
-package net.openhft.chronicle.wire.internal;
+package net.openhft.chronicle.wire;
 
 import net.openhft.chronicle.wire.MarshallableOut;
 import net.openhft.chronicle.wire.WireType;
+import net.openhft.chronicle.wire.internal.StringConsumerMarshallableOut;
 import org.junit.Test;
 
 import java.io.StringWriter;
