+++ src/test/java/net/openhft/chronicle/wire/WireTypeConverterTest.java
@@ -1,6 +1,5 @@
-package net.openhft.chronicle.wire.internal;
+package net.openhft.chronicle.wire;
 
-import net.openhft.chronicle.wire.WireTypeConverter;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -8,13 +7,15 @@ import static junit.framework.TestCase.assertEquals;
 
 public class WireTypeConverterTest {
 
-    private final String json = "{\"@net.openhft.chronicle.wire.internal.MyClass\":{\"msg\":\"msg:\\\"hello\\\"\"}}";
-    private final String yaml = "!net.openhft.chronicle.wire.internal.MyClass {\n" +
+    private final String json = "!net.openhft.chronicle.wire.MyClass {\n" +
+            "  msg: \"msg:\\\"hello\\\"\"\n" +
+            "}\n";
+    private final String yaml = "!net.openhft.chronicle.wire.MyClass {\n" +
             "  msg: \"msg:\\\"hello\\\"\"\n" +
             "}\n";
 
-    private final String jsonUnknownClass = "{\"@net.openhft.chronicle.wire.internal.UnknownClass\":{\"msg\":\"msg:\\\"hello\\\"\"}}";
-    private final String yamlUnknownClass = "!net.openhft.chronicle.wire.internal.UnknownClass {\n" +
+    private final String jsonUnknownClass = "{\"@net.openhft.chronicle.wire.UnknownClass\":{\"msg\":\"msg:\\\"hello\\\"\"}}";
+    private final String yamlUnknownClass = "!net.openhft.chronicle.wire.UnknownClass {\n" +
             "  msg: \"msg:\\\"hello\\\"\"\n" +
             "}\n";
 
@@ -45,27 +46,30 @@ public class WireTypeConverterTest {
 
     @Test(expected = ClassCastException.class)
     public void testYamlClassCastException() throws Exception {
-        new WireTypeConverter().yamlToJson("!net.openhft.chronicle.wire.internal.MyClass2 {\n" +
-                "  myClass: !net.openhft.chronicle.wire.internal.MyClass2 { x: aa }\n" +
+        new WireTypeConverter().yamlToJson("!net.openhft.chronicle.wire.MyClass2 {\n" +
+                "  myClass: !net.openhft.chronicle.wire.MyClass2 { x: aa }\n" +
                 "}\n");
     }
 
     @Test
     public void testYamlNoClassCastException() throws Exception {
         final WireTypeConverter converter = new WireTypeConverter();
-        converter.addAlias(MyClass3.class, "net.openhft.chronicle.wire.internal.MyOldClass");
-        final CharSequence json = converter.yamlToJson("!net.openhft.chronicle.wire.internal.MyClass2 {\n" +
-                "  myClass: !net.openhft.chronicle.wire.internal.MyOldClass { x: abc }\n" +
+        converter.addAlias(MyClass3.class, "net.openhft.chronicle.wire.MyOldClass");
+        final CharSequence json = converter.yamlToJson("!net.openhft.chronicle.wire.MyClass2 {\n" +
+                "  myClass: !net.openhft.chronicle.wire.MyOldClass { x: abc }\n" +
                 "}\n");
-        assertEquals("" +
-                        "{\"@net.openhft.chronicle.wire.internal.MyClass2\":{\"myClass\":{\"@MyOldClass\":{\"x\":\"abc\"}}}}",
+        assertEquals("!net.openhft.chronicle.wire.MyClass2 {\n" +
+                        "  myClass: {\n" +
+                        "    x: abc\n" +
+                        "  }\n" +
+                        "}\n",
                 json.toString());
     }
 
     @Test(expected = ClassCastException.class)
     public void testJsonClassCastException() throws Exception {
-        new WireTypeConverter().jsonToYaml("{\"@net.openhft.chronicle.wire.internal.MyClass2\": {\n" +
-                " myClass: {\"@net.openhft.chronicle.wire.internal.MyClass2\": { \"x\": \"bb\" } }\n" +
+        new WireTypeConverter().jsonToYaml("{\"@net.openhft.chronicle.wire.MyClass2\": {\n" +
+                " myClass: {\"@net.openhft.chronicle.wire.MyClass2\": { \"x\": \"bb\" } }\n" +
                 "}\n");
     }
 
@@ -74,11 +78,11 @@ public class WireTypeConverterTest {
         final WireTypeConverter converter = new WireTypeConverter();
         converter.addAlias(MyClass3.class, "MyOldClass");
         final CharSequence yaml = converter.jsonToYaml("" +
-                "{\"@net.openhft.chronicle.wire.internal.MyClass2\": {\n" +
+                "{\"@net.openhft.chronicle.wire.MyClass2\": {\n" +
                 "  myClass: {\"@MyOldClass\": { x: abcd } }\n" +
                 "}\n");
         assertEquals("" +
-                "!net.openhft.chronicle.wire.internal.MyClass2 {\n" +
+                "!net.openhft.chronicle.wire.MyClass2 {\n" +
                 "  myClass: {\n" +
                 "    x: abcd\n" +
                 "  }\n" +
