--- src/test/java/net/openhft/chronicle/wire/internal/MyClass2.java
@@ -1,9 +1,0 @@
-package net.openhft.chronicle.wire.internal;
-
-import net.openhft.chronicle.wire.SelfDescribingMarshallable;
-
-public class MyClass2 extends SelfDescribingMarshallable {
-
-    MyClass3 myClass;
-
-}
